Update the OWIN/Katana bits to use the vNext Data Protection stack to encrypt the RSA keys and improve the keys directory detection logic
