+++ src/AspNet.Security.OpenIdConnect.Server/OpenIdConnectServerHelpers.cs
@@ -154,7 +154,11 @@ namespace AspNet.Security.OpenIdConnect.Server {
             if (!string.IsNullOrEmpty(Environment.GetEnvironmentVariable("WEBSITE_INSTANCE_ID"))) {
                 path = Environment.GetEnvironmentVariable("HOME");
                 if (!string.IsNullOrEmpty(path)) {
-                    return GetKeyStorageDirectoryFromBasePath(path);
+                    try {
+                        return GetKeyStorageDirectoryFromBasePath(path);
+                    }
+
+                    catch { }
                 }
             }
 
@@ -163,7 +167,11 @@ namespace AspNet.Security.OpenIdConnect.Server {
             path = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData);
 
             if (!string.IsNullOrEmpty(path)) {
-                return GetKeyStorageDirectoryFromBasePath(path);
+                try {
+                    return GetKeyStorageDirectoryFromBasePath(path);
+                }
+
+                catch { }
             }
 #else
 
@@ -171,41 +179,61 @@ namespace AspNet.Security.OpenIdConnect.Server {
             // using the LOCALAPPDATA environment variable.
             path = Environment.GetEnvironmentVariable("LOCALAPPDATA");
             if (!string.IsNullOrEmpty(path)) {
-                return GetKeyStorageDirectoryFromBasePath(path);
+                try {
+                    return GetKeyStorageDirectoryFromBasePath(path);
+                }
+
+                catch { }
             }
 
             // If the LOCALAPPDATA environment variable was not found,
             // try to determine the actual AppData/Local path from USERPROFILE.
             path = Environment.GetEnvironmentVariable("USERPROFILE");
             if (!string.IsNullOrEmpty(path)) {
-                return GetKeyStorageDirectoryFromBasePath(Path.Combine(path, "AppData", "Local"));
+                try {
+                    return GetKeyStorageDirectoryFromBasePath(Path.Combine(path, "AppData", "Local"));
+                }
+
+                catch { }
             }
 
             // On Linux environments, use the HOME variable.
             path = Environment.GetEnvironmentVariable("HOME");
             if (!string.IsNullOrEmpty(path)) {
-                return new DirectoryInfo(Path.Combine(path, ".aspnet", "aspnet-contrib", "aspnet-oidc-server"));
+                try {
+                    return Directory.CreateDirectory(Path.Combine(path, ".aspnet", "aspnet-contrib", "oidc-server"));
+                }
+
+                catch { }
             }
 #endif
 
             // Use the ASPNET_TEMP environment variable when specified.
             path = Environment.GetEnvironmentVariable("ASPNET_TEMP");
             if (!string.IsNullOrEmpty(path)) {
-                return GetKeyStorageDirectoryFromBasePath(path);
+                try {
+                    return GetKeyStorageDirectoryFromBasePath(path);
+                }
+
+                catch { }
             }
 
             // Note: returning the TEMP directory is safe as keys are always encrypted using the
             // data protection system, making the keys unreadable outside this environment.
             path = Path.GetTempPath();
             if (!string.IsNullOrEmpty(path)) {
-                return GetKeyStorageDirectoryFromBasePath(path);
+                try {
+                    return GetKeyStorageDirectoryFromBasePath(path);
+                }
+
+                catch { }
             }
 
             throw new InvalidOperationException("No directory can be found to store the RSA keys.");
         }
 
         private static DirectoryInfo GetKeyStorageDirectoryFromBasePath(string path) {
-            return new DirectoryInfo(Path.Combine(path, "ASP.NET", "aspnet-contrib", "aspnet-oidc-server"));
+            return Directory.CreateDirectory(Path.Combine(path, "ASP.NET", "aspnet-contrib", "oidc-server"));
         }
 
         internal static string GetIssuer(this HttpContext context, OpenIdConnectServerOptions options) {
