+++ src/AspNet.Security.OpenIdConnect.Server/OpenIdConnectServerMiddleware.cs
@@ -5,14 +5,19 @@
  */
 
 using System;
+using System.Diagnostics;
+using System.IO;
 using System.Text.Encodings.Web;
 using JetBrains.Annotations;
 using Microsoft.AspNetCore.Authentication;
+using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.DataProtection;
+using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Caching.Distributed;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using Microsoft.IdentityModel.Tokens;
 
 namespace AspNet.Security.OpenIdConnect.Server {
     /// <summary>
@@ -29,6 +34,7 @@ namespace AspNet.Security.OpenIdConnect.Server {
         public OpenIdConnectServerMiddleware(
             [NotNull] RequestDelegate next,
             [NotNull] IOptions<OpenIdConnectServerOptions> options,
+            [NotNull] IHostingEnvironment environment,
             [NotNull] ILoggerFactory loggerFactory,
             [NotNull] IDistributedCache cache,
             [NotNull] HtmlEncoder htmlEncoder,
@@ -106,6 +112,75 @@ namespace AspNet.Security.OpenIdConnect.Server {
 
             if (Options.HtmlEncoder == null) {
                 Options.HtmlEncoder = htmlEncoder;
+            }
+
+            // If no key has been explicitly added, use the fallback mode.
+            if (Options.SigningCredentials.Count == 0) {
+                if (environment.IsProduction()) {
+                    Logger.LogWarning("No explicit signing credentials have been registered. " +
+                                      "Using a X.509 certificate stored in the machine store " +
+                                      "is recommended for production environments.");
+                }
+
+                var directory = OpenIdConnectServerHelpers.GetDefaultKeyStorageDirectory();
+                Debug.Assert(directory != null);
+
+                // Get a data protector from the services provider.
+                var protector = dataProtectionProvider.CreateProtector(
+                    nameof(OpenIdConnectServerMiddleware),
+                    Options.AuthenticationScheme, "Keys", "v1");
+
+                foreach (var file in directory.EnumerateFiles("*.key")) {
+                    using (var buffer = new MemoryStream())
+                    using (var stream = file.Open(FileMode.Open, FileAccess.Read, FileShare.Read)) {
+                        // Copy the key content to the buffer.
+                        stream.CopyTo(buffer);
+
+                        // Extract the key material using the data protector.
+                        // Ignore the key if the decryption process failed.
+                        string usage;
+                        var parameters = OpenIdConnectServerHelpers.DecryptKey(protector, buffer.ToArray(), out usage);
+                        if (parameters == null) {
+                            Logger.LogDebug("An invalid/incompatible key was ignored: {Key}.", file.FullName);
+
+                            continue;
+                        }
+
+                        // Only add the key if it corresponds to the intended usage.
+                        if (!string.Equals(usage, "Signing", StringComparison.OrdinalIgnoreCase)) {
+                            continue;
+                        }
+
+                        Logger.LogInformation("An existing key was automatically added to the " +
+                                              "signing credentials list: {Key}.", file.FullName);
+
+                        // Add the key to the signing credentials list.
+                        Options.SigningCredentials.AddKey(new RsaSecurityKey(parameters.Value));
+                    }
+                }
+
+                // If no signing key has been found, generate and persist a new RSA key.
+                if (Options.SigningCredentials.Count == 0) {
+                    // Generate a new RSA key and export its public/private parameters.
+                    var provider = OpenIdConnectServerHelpers.GenerateKey(size: 2048);
+                    var parameters = provider.ExportParameters(/* includePrivateParameters */ true);
+
+                    // Generate a new file name for the key and determine its absolute path.
+                    var path = Path.Combine(directory.FullName, Guid.NewGuid().ToString() + ".key");
+
+                    using (var stream = new FileStream(path, FileMode.CreateNew, FileAccess.Write)) {
+                        // Encrypt the key using the data protector.
+                        var bytes = OpenIdConnectServerHelpers.EncryptKey(protector, parameters, usage: "Signing");
+
+                        // Write the encrypted key to the file stream.
+                        stream.Write(bytes, 0, bytes.Length);
+                    }
+
+                    Logger.LogInformation("A new RSA key was automatically generated, added to the " +
+                                          "signing credentials list and persisted on the disk: {Path}.", path);
+
+                    Options.SigningCredentials.AddKey(new RsaSecurityKey(parameters));
+                }
             }
         }
