+++ src/Owin.Security.OpenIdConnect.Server/OpenIdConnectServerHelpers.cs
@@ -4,8 +4,8 @@ using System.IO;
 using System.Linq;
 using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
+using Microsoft.AspNetCore.DataProtection;
 using Microsoft.Owin;
-using Microsoft.Owin.Security.DataProtection;
 
 namespace Owin.Security.OpenIdConnect.Server {
     internal static class OpenIdConnectServerHelpers {
@@ -146,7 +146,11 @@ namespace Owin.Security.OpenIdConnect.Server {
             if (!string.IsNullOrEmpty(Environment.GetEnvironmentVariable("WEBSITE_INSTANCE_ID"))) {
                 path = Environment.GetEnvironmentVariable("HOME");
                 if (!string.IsNullOrEmpty(path)) {
-                    return GetKeyStorageDirectoryFromBasePath(path);
+                    try {
+                        return GetKeyStorageDirectoryFromBasePath(path);
+                    }
+
+                    catch { }
                 }
             }
 
@@ -154,47 +158,49 @@ namespace Owin.Security.OpenIdConnect.Server {
             path = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData);
 
             if (!string.IsNullOrEmpty(path)) {
-                return GetKeyStorageDirectoryFromBasePath(path);
-            }
+                try {
+                    return GetKeyStorageDirectoryFromBasePath(path);
+                }
 
-            // Try to resolve the AppData/Local folder
-            // using the LOCALAPPDATA environment variable.
-            path = Environment.GetEnvironmentVariable("LOCALAPPDATA");
-            if (!string.IsNullOrEmpty(path)) {
-                return GetKeyStorageDirectoryFromBasePath(path);
-            }
-
-            // If the LOCALAPPDATA environment variable was not found,
-            // try to determine the actual AppData/Local path from USERPROFILE.
-            path = Environment.GetEnvironmentVariable("USERPROFILE");
-            if (!string.IsNullOrEmpty(path)) {
-                return GetKeyStorageDirectoryFromBasePath(Path.Combine(path, "AppData", "Local"));
+                catch { }
             }
 
             // On Linux environments, use the HOME variable.
             path = Environment.GetEnvironmentVariable("HOME");
             if (!string.IsNullOrEmpty(path)) {
-                return new DirectoryInfo(Path.Combine(path, ".aspnet", "aspnet-contrib", "owin-oidc-server"));
+                try {
+                    return Directory.CreateDirectory(Path.Combine(path, ".aspnet", "aspnet-contrib", "oidc-server"));
+                }
+
+                catch { }
             }
 
             // Use the ASPNET_TEMP environment variable when specified.
             path = Environment.GetEnvironmentVariable("ASPNET_TEMP");
             if (!string.IsNullOrEmpty(path)) {
-                return GetKeyStorageDirectoryFromBasePath(path);
+                try {
+                    return GetKeyStorageDirectoryFromBasePath(path);
+                }
+
+                catch { }
             }
 
             // Note: returning the TEMP directory is safe as keys are always encrypted using the
             // data protection system, making the keys unreadable outside this environment.
             path = Path.GetTempPath();
             if (!string.IsNullOrEmpty(path)) {
-                return GetKeyStorageDirectoryFromBasePath(path);
+                try {
+                    return GetKeyStorageDirectoryFromBasePath(path);
+                }
+
+                catch { }
             }
 
             throw new InvalidOperationException("No directory can be found to store the RSA keys.");
         }
 
         private static DirectoryInfo GetKeyStorageDirectoryFromBasePath(string path) {
-            return new DirectoryInfo(Path.Combine(path, "ASP.NET", "aspnet-contrib", "owin-oidc-server"));
+            return Directory.CreateDirectory(Path.Combine(path, "ASP.NET", "aspnet-contrib", "oidc-server"));
         }
 
         internal static string GetIssuer(this IOwinContext context, OpenIdConnectServerOptions options) {
