+++ src/Owin.Security.OpenIdConnect.Server/OpenIdConnectServerMiddleware.cs
@@ -5,6 +5,10 @@
  */
 
 using System;
+using System.Diagnostics;
+using System.IdentityModel.Tokens;
+using System.IO;
+using System.Security.Cryptography;
 using Microsoft.AspNetCore.DataProtection;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
@@ -111,6 +115,113 @@ namespace Owin.Security.OpenIdConnect.Server {
                     Options.AuthenticationType, "Refresh_Token", "v1");
 
                 Options.RefreshTokenFormat = new AspNetTicketDataFormat(new DataProtectorShim(protector));
+            }
+
+            // If no key has been explicitly added, use the fallback mode.
+            if (Options.EncryptingCredentials.Count == 0 || Options.SigningCredentials.Count == 0) {
+                var directory = OpenIdConnectServerHelpers.GetDefaultKeyStorageDirectory();
+                Debug.Assert(directory != null);
+
+                // Get a data protector from the services provider.
+                var protector = Options.DataProtectionProvider.CreateProtector(
+                    nameof(OpenIdConnectServerMiddleware),
+                    Options.AuthenticationType, "Keys", "v1");
+
+                foreach (var file in directory.EnumerateFiles("*.key")) {
+                    using (var buffer = new MemoryStream())
+                    using (var stream = file.Open(FileMode.Open, FileAccess.Read, FileShare.Read)) {
+                        // Copy the key content to the buffer.
+                        stream.CopyTo(buffer);
+
+                        // Extract the key material using the data protector.
+                        // Ignore the key if the decryption process failed.
+                        string usage;
+                        var parameters = OpenIdConnectServerHelpers.DecryptKey(protector, buffer.ToArray(), out usage);
+                        if (parameters == null) {
+                            Options.Logger.LogDebug("An invalid/incompatible key was ignored: {Key}.", file.FullName);
+
+                            continue;
+                        }
+
+                        if (string.Equals(usage, "Encryption", StringComparison.OrdinalIgnoreCase)) {
+                            // Only add the encryption key if none has been explictly added.
+                            if (Options.EncryptingCredentials.Count != 0) {
+                                continue;
+                            }
+
+                            var algorithm = RSA.Create();
+                            algorithm.ImportParameters(parameters.Value);
+
+                            // Add the key to the encryption credentials list.
+                            Options.EncryptingCredentials.AddKey(new RsaSecurityKey(algorithm));
+
+                            Options.Logger.LogInformation("An existing key was automatically added to the " +
+                                                          "encryption credentials list: {Key}.", file.FullName);
+                        }
+
+                        else if (string.Equals(usage, "Signing", StringComparison.OrdinalIgnoreCase)) {
+                            // Only add the signing key if none has been explictly added.
+                            if (Options.SigningCredentials.Count != 0) {
+                                continue;
+                            }
+
+                            var algorithm = RSA.Create();
+                            algorithm.ImportParameters(parameters.Value);
+
+                            // Add the key to the signing credentials list.
+                            Options.SigningCredentials.AddKey(new RsaSecurityKey(algorithm));
+
+                            Options.Logger.LogInformation("An existing key was automatically added to the " +
+                                                          "signing credentials list: {Key}.", file.FullName);
+                        }
+                    }
+                }
+
+                // If no encryption key has been found, generate and persist a new RSA key.
+                if (Options.EncryptingCredentials.Count == 0) {
+                    // Generate a new RSA key and export its public/private parameters.
+                    var algorithm = OpenIdConnectServerHelpers.GenerateKey(size: 2048);
+                    var parameters = algorithm.ExportParameters(/* includePrivateParameters: */ true);
+
+                    // Generate a new file name for the key and determine its absolute path.
+                    var path = Path.Combine(directory.FullName, Guid.NewGuid().ToString() + ".key");
+
+                    using (var stream = new FileStream(path, FileMode.CreateNew, FileAccess.Write)) {
+                        // Encrypt the key using the data protector.
+                        var bytes = OpenIdConnectServerHelpers.EncryptKey(protector, parameters, usage: "Encryption");
+
+                        // Write the encrypted key to the file stream.
+                        stream.Write(bytes, 0, bytes.Length);
+                    }
+
+                    Options.EncryptingCredentials.AddKey(new RsaSecurityKey(algorithm));
+
+                    Options.Logger.LogInformation("A new RSA key was automatically generated, added to the " +
+                                                  "encryption credentials list and persisted on the disk: {Path}.", path);
+                }
+
+                // If no signing key has been found, generate and persist a new RSA key.
+                if (Options.SigningCredentials.Count == 0) {
+                    // Generate a new RSA key and export its public/private parameters.
+                    var algorithm = OpenIdConnectServerHelpers.GenerateKey(size: 2048);
+                    var parameters = algorithm.ExportParameters(/* includePrivateParameters: */ true);
+
+                    // Generate a new file name for the key and determine its absolute path.
+                    var path = Path.Combine(directory.FullName, Guid.NewGuid().ToString() + ".key");
+
+                    using (var stream = new FileStream(path, FileMode.CreateNew, FileAccess.Write)) {
+                        // Encrypt the key using the data protector.
+                        var bytes = OpenIdConnectServerHelpers.EncryptKey(protector, parameters, usage: "Signing");
+
+                        // Write the encrypted key to the file stream.
+                        stream.Write(bytes, 0, bytes.Length);
+                    }
+
+                    Options.SigningCredentials.AddKey(new RsaSecurityKey(algorithm));
+
+                    Options.Logger.LogInformation("A new RSA key was automatically generated, added to the " +
+                                                  "signing credentials list and persisted on the disk: {Path}.", path);
+                }
             }
         }
